network_information_source/requirements.txt
dnspython==2.2.1
dpcontracts==0.6.0
fake_useragent==1.1.3
kafka_python3==3.0.0
langid==1.1.6
python_whois==0.8.0
easy_spider_tool==1.0.7
sshtunnel==0.4.0
tldextract==3.1.2

